adding example script for cli functions
