+++ ochre/cli.py
@@ -1,5 +1,4 @@
 import os
-import shutil
 import functools
 import time
 import datetime as dt
@@ -8,15 +7,9 @@ from multiprocessing import Pool
 import click
 
 from ochre import Dwelling, Analysis
-from ochre.utils import default_input_path
-
 
 # Functions for command line interface (CLI). Uses a limited number of options
 # for running OCHRE
-
-# Note: see documentation for where to download other weather files
-# https://ochre-nrel.readthedocs.io/en/latest/InputsAndArguments.html#weather-file
-default_weather_file = os.path.join(default_input_path, "Weather", "G0800310.epw")
 
 
 def update_output_path(output_path, input_path):
@@ -205,89 +198,10 @@ def run_multiple_local(
     my_print("All processes finished, exiting.")
 
 
-def compile_results(main_path, n_max=None):
-    # Sample script to compile results from multiple OCHRE runs
-
-    # set up folder for compiled results
-    output_path = os.path.join(main_path, "compiled")
-    os.makedirs(output_path, exist_ok=True)
-    my_print("Compiling OCHRE results for:", main_path)
-
-    # find all building folders
-    required_files = ["ochre_complete"]
-    run_paths = Analysis.find_subfolders(main_path, required_files)
-    n = len(run_paths)
-
-    # ensure at least 1 run folder found
-    if not n:
-        my_print("No buildings found in:", main_path)
-        return
-    else:
-        my_print(f"Found {n} completed simulations in:", main_path)
-
-    # limit total number of runs
-    if n_max is not None and n > n_max:
-        my_print(f"Limiting number of runs to {n_max}")
-        run_paths = run_paths[:n_max]
-        n = n_max
-
-    run_names = {os.path.relpath(path, main_path): path for path in run_paths}
-
-    # combine input json files
-    json_files = {name: os.path.join(path, "ochre.json") for name, path in run_names.items()}
-    df = Analysis.combine_json_files(json_files)
-    df.to_csv(os.path.join(output_path, "all_ochre_inputs.csv"))
-
-    # combine metrics files
-    metrics_files = {name: os.path.join(path, "ochre_metrics.csv") for name, path in run_names.items()}
-    df = Analysis.combine_metrics_files(metrics_files)
-    df.to_csv(os.path.join(output_path, "all_ochre_metrics.csv"))
-
-    # combine single time series column for each house (e.g., total electricity consumption)
-    results_files = {name: os.path.join(path, "ochre.csv") for name, path in run_names.items()}
-    df = Analysis.combine_time_series_column("Total Electric Power (kW)", results_files)
-    df.to_csv(os.path.join(output_path, "all_ochre_total_powers.csv"))
-
-    # aggregate time series data across all simulations
-    df = Analysis.combine_time_series_files(results_files, agg_type="House")
-    df.to_csv(os.path.join(output_path, "all_ochre_results.csv"))
-
-    my_print("Saved OCHRE results to:", output_path)
-
-
 def my_print(*args):
     # prints with date and other info
     now = dt.datetime.now()
     print(now, *args)
-
-
-if __name__ == "__main__":
-    main_path = os.getcwd()
-
-    # Download ResStock files to current directory
-    buildings = ["bldg0112631"]
-    upgrades = ["up00", "up11"]
-    input_paths = []
-    for upgrade in upgrades:
-        for building in buildings:
-            input_path = os.path.join(main_path, building, upgrade)
-            os.makedirs(input_path, exist_ok=True)
-            Analysis.download_resstock_model(building, upgrade, input_path, overwrite=False)
-            shutil.copy(default_weather_file, input_path)
-            input_paths.append(input_path)
-
-    # Run Dwelling models sequentially
-    # for input_path in input_paths:
-    #     run_single_building(input_path, duration=7)
-
-    # Run simulations in parallel
-    run_multiple_local(main_path, n_parallel=2, overwrite=True, duration=7)
-
-    # Run simulations on HPC using Slurm
-    # run_multiple_hpc(main_path, overwrite=True)
-
-    # Compile simulation data
-    compile_results(main_path)
 
 
 @click.group()
